Add test to create PyTorchJob from func

+++ sdk/python/test/e2e/test_e2e_pytorchjob.py
@@ -36,7 +36,7 @@ from test.e2e.constants import TEST_GANG_SCHEDULER_NAME_ENV_KEY
 from test.e2e.constants import GANG_SCHEDULERS, NONE_GANG_SCHEDULERS
 
 logging.basicConfig(format="%(message)s")
-logging.getLogger().setLevel(logging.INFO)
+logging.getLogger("kubeflow.training.api.training_client").setLevel(logging.DEBUG)
 
 TRAINING_CLIENT = TrainingClient(job_kind=constants.PYTORCHJOB_KIND)
 JOB_NAME = "pytorchjob-mnist-ci-test"
@@ -156,6 +156,45 @@ def test_sdk_e2e(job_namespace):
         utils.print_job_results(TRAINING_CLIENT, JOB_NAME, job_namespace)
         TRAINING_CLIENT.delete_job(JOB_NAME, job_namespace)
         raise Exception(f"PyTorchJob E2E fails. Exception: {e}")
+
+    utils.print_job_results(TRAINING_CLIENT, JOB_NAME, job_namespace)
+    TRAINING_CLIENT.delete_job(JOB_NAME, job_namespace)
+
+
+def test_sdk_e2e_create_from_func(job_namespace):
+    def train_func():
+        import time
+
+        for i in range(10):
+            print(f"Start training for Epoch {i}")
+            time.sleep(1)
+
+    num_workers = 1
+
+    TRAINING_CLIENT.create_job(
+        name=JOB_NAME,
+        namespace=job_namespace,
+        train_func=train_func,
+        num_worker_replicas=num_workers,
+    )
+
+    logging.info(f"List of created {TRAINING_CLIENT.job_kind}s")
+    logging.info(TRAINING_CLIENT.list_jobs(job_namespace))
+
+    try:
+        utils.verify_job_e2e(TRAINING_CLIENT, JOB_NAME, job_namespace, wait_timeout=900)
+    except Exception as e:
+        utils.print_job_results(TRAINING_CLIENT, JOB_NAME, job_namespace)
+        TRAINING_CLIENT.delete_job(JOB_NAME, job_namespace)
+        raise Exception(f"PyTorchJob create from function E2E fails. Exception: {e}")
+
+    # Verify that PyTorchJob has correct pods.
+    pod_names = TRAINING_CLIENT.get_job_pod_names(
+        name=JOB_NAME, namespace=job_namespace
+    )
+
+    if len(pod_names) != num_workers or f"{JOB_NAME}-worker-0" not in pod_names:
+        raise Exception(f"PyTorchJob has incorrect pods: {pod_names}")
 
     utils.print_job_results(TRAINING_CLIENT, JOB_NAME, job_namespace)
     TRAINING_CLIENT.delete_job(JOB_NAME, job_namespace)
